student-report-synthesis/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
openai>=1.0.0
python-dotenv==1.0.0

# Report generation
jinja2==3.1.2
reportlab==3.6.12
pillow==9.5.0
python-docx==0.8.11

# PDF conversion options
xhtml2pdf==0.2.11
weasyprint>=53.0
beautifulsoup4>=4.9.3

# Image processing
requests>=2.28.0

# Data handling
numpy>=1.22.0
pandas>=1.3.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
flake8>=4.0.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0